// sdk/ts/src/generated/protos/cosmos/auth/v1beta1/query.ts
export interface QueryAccountsRequest {
  /** pagination defines an optional pagination for the request. */
  pagination: PageRequest | undefined;
}

/** QueryAccountsResponse is the response type for the Query/Accounts RPC method. */
export interface QueryAccountsResponse {
  /** accounts are the existing accounts */
  accounts: Any[];
  /** pagination defines the pagination in the response. */
  pagination: PageResponse | undefined;
}

/** QueryAccountRequest is the request type for the Query/Account RPC method. */
export interface QueryAccountRequest {
  /** address defines the address to query for. */
  address: string;
}

/** QueryAccountResponse is the response type for the Query/Account RPC method. */
export interface QueryAccountResponse {
  /** account defines the account of the corresponding address. */
  account: Any | undefined;
}

/** QueryParamsRequest is the request type for the Query/Params RPC method. */
export interface QueryParamsRequest {
}

/** QueryParamsResponse is the response type for the Query/Params RPC method. */
export interface QueryParamsResponse {
  /** params defines the parameters of the module. */
  params: Params | undefined;
}

/** QueryModuleAccountsRequest is the request type for the Query/ModuleAccounts RPC method. */
export interface QueryModuleAccountsRequest {
}

/** QueryModuleAccountsResponse is the response type for the Query/ModuleAccounts RPC method. */
export interface QueryModuleAccountsResponse {
  accounts: Any[];
}

/** QueryModuleAccountByNameRequest is the request type for the Query/ModuleAccountByName RPC method. */
export interface QueryModuleAccountByNameRequest {
  name: string;
}

/** QueryModuleAccountByNameResponse is the response type for the Query/ModuleAccountByName RPC method. */
export interface QueryModuleAccountByNameResponse {
  account: Any | undefined;
}

/** Bech32PrefixRequest is the request type for Bech32Prefix rpc method. */
export interface Bech32PrefixRequest {
}

/** Bech32PrefixResponse is the response type for Bech32Prefix rpc method. */
export interface Bech32PrefixResponse {
  bech32Prefix: string;
}

/** AddressBytesToStringRequest is the request type for AddressString rpc method. */
export interface AddressBytesToStringRequest {
  addressBytes: Uint8Array;
}

/** AddressBytesToStringResponse is the response type for AddressString rpc method. */
export interface AddressBytesToStringResponse {
  addressString: string;
}

/** AddressStringToBytesRequest is the request type for AccountBytes rpc method. */
export interface AddressStringToBytesRequest {
  addressString: string;
}

/** AddressStringToBytesResponse is the response type for AddressBytes rpc method. */
export interface AddressStringToBytesResponse {
  addressBytes: Uint8Array;
}

/** QueryAccountAddressByIDRequest is the request type for AccountAddressByID rpc method */
export interface QueryAccountAddressByIDRequest {
  /**
   * Deprecated, use account_id instead
   *
   * id is the account number of the address to be queried. This field
   * should have been an uint64 (like all account numbers), and will be
   * updated to uint64 in a future version of the auth query.
   *
   * @deprecated
   */
  id: Long;
  /** account_id is the account number of the address to be queried. */
  accountId: Long;
}

/** QueryAccountAddressByIDResponse is the response type for AccountAddressByID rpc method */
export interface QueryAccountAddressByIDResponse {
  accountAddress: string;
}

/** QueryAccountInfoRequest is the Query/AccountInfo request type. */
export interface QueryAccountInfoRequest {
  /** address is the account address string. */
  address: string;
}

/** QueryAccountInfoResponse is the Query/AccountInfo response type. */
export interface QueryAccountInfoResponse {
  /** info is the account info which is represented by BaseAccount. */
  info: BaseAccount | undefined;
}
// ...
export const Bech32PrefixRequest: MessageFns<Bech32PrefixRequest, "cosmos.auth.v1beta1.Bech32PrefixRequest"> = {
  $type: "cosmos.auth.v1beta1.Bech32PrefixRequest" as const,

  encode(_: Bech32PrefixRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Bech32PrefixRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseBech32PrefixRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(_: any): Bech32PrefixRequest {
    return {};
  },

  toJSON(_: Bech32PrefixRequest): unknown {
    const obj: any = {};
    return obj;
  },
  fromPartial(_: DeepPartial<Bech32PrefixRequest>): Bech32PrefixRequest {
    const message = createBaseBech32PrefixRequest();
    return message;
  },
};

function createBaseBech32PrefixResponse(): Bech32PrefixResponse {
  return { bech32Prefix: "" };
}

export const Bech32PrefixResponse: MessageFns<Bech32PrefixResponse, "cosmos.auth.v1beta1.Bech32PrefixResponse"> = {
  $type: "cosmos.auth.v1beta1.Bech32PrefixResponse" as const,

  encode(message: Bech32PrefixResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.bech32Prefix !== "") {
      writer.uint32(10).string(message.bech32Prefix);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Bech32PrefixResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseBech32PrefixResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.bech32Prefix = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Bech32PrefixResponse {
    return { bech32Prefix: isSet(object.bech32_prefix) ? globalThis.String(object.bech32_prefix) : "" };
  },

  toJSON(message: Bech32PrefixResponse): unknown {
    const obj: any = {};
    if (message.bech32Prefix !== "") {
      obj.bech32_prefix = message.bech32Prefix;
    }
    return obj;
  },
  fromPartial(object: DeepPartial<Bech32PrefixResponse>): Bech32PrefixResponse {
    const message = createBaseBech32PrefixResponse();
    message.bech32Prefix = object.bech32Prefix ?? "";
    return message;
  },
};

function createBaseAddressBytesToStringRequest(): AddressBytesToStringRequest {
  return { addressBytes: new Uint8Array(0) };
}

export const AddressBytesToStringRequest: MessageFns<
  AddressBytesToStringRequest,
  "cosmos.auth.v1beta1.AddressBytesToStringRequest"
> = {
  $type: "cosmos.auth.v1beta1.AddressBytesToStringRequest" as const,

  encode(message: AddressBytesToStringRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.addressBytes.length !== 0) {
      writer.uint32(10).bytes(message.addressBytes);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AddressBytesToStringRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAddressBytesToStringRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.addressBytes = reader.bytes();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AddressBytesToStringRequest {
    return { addressBytes: isSet(object.address_bytes) ? bytesFromBase64(object.address_bytes) : new Uint8Array(0) };
  },

  toJSON(message: AddressBytesToStringRequest): unknown {
    const obj: any = {};
    if (message.addressBytes.length !== 0) {
      obj.address_bytes = base64FromBytes(message.addressBytes);
    }
    return obj;
  },
  fromPartial(object: DeepPartial<AddressBytesToStringRequest>): AddressBytesToStringRequest {
    const message = createBaseAddressBytesToStringRequest();
    message.addressBytes = object.addressBytes ?? new Uint8Array(0);
    return message;
  },
};

function createBaseAddressBytesToStringResponse(): AddressBytesToStringResponse {
  return { addressString: "" };
}

export const AddressBytesToStringResponse: MessageFns<
  AddressBytesToStringResponse,
  "cosmos.auth.v1beta1.AddressBytesToStringResponse"
> = {
  $type: "cosmos.auth.v1beta1.AddressBytesToStringResponse" as const,

  encode(message: AddressBytesToStringResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.addressString !== "") {
      writer.uint32(10).string(message.addressString);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AddressBytesToStringResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAddressBytesToStringResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.addressString = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AddressBytesToStringResponse {
    return { addressString: isSet(object.address_string) ? globalThis.String(object.address_string) : "" };
  },

  toJSON(message: AddressBytesToStringResponse): unknown {
    const obj: any = {};
    if (message.addressString !== "") {
      obj.address_string = message.addressString;
    }
    return obj;
  },
  fromPartial(object: DeepPartial<AddressBytesToStringResponse>): AddressBytesToStringResponse {
    const message = createBaseAddressBytesToStringResponse();
    message.addressString = object.addressString ?? "";
    return message;
  },
};

function createBaseAddressStringToBytesRequest(): AddressStringToBytesRequest {
  return { addressString: "" };
}

export const AddressStringToBytesRequest: MessageFns<
  AddressStringToBytesRequest,
  "cosmos.auth.v1beta1.AddressStringToBytesRequest"
> = {
  $type: "cosmos.auth.v1beta1.AddressStringToBytesRequest" as const,

  encode(message: AddressStringToBytesRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.addressString !== "") {
      writer.uint32(10).string(message.addressString);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AddressStringToBytesRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAddressStringToBytesRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.addressString = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AddressStringToBytesRequest {
    return { addressString: isSet(object.address_string) ? globalThis.String(object.address_string) : "" };
  },

  toJSON(message: AddressStringToBytesRequest): unknown {
    const obj: any = {};
    if (message.addressString !== "") {
      obj.address_string = message.addressString;
    }
    return obj;
  },
  fromPartial(object: DeepPartial<AddressStringToBytesRequest>): AddressStringToBytesRequest {
    const message = createBaseAddressStringToBytesRequest();
    message.addressString = object.addressString ?? "";
    return message;
  },
};

function createBaseAddressStringToBytesResponse(): AddressStringToBytesResponse {
  return { addressBytes: new Uint8Array(0) };
}

export const AddressStringToBytesResponse: MessageFns<
  AddressStringToBytesResponse,
  "cosmos.auth.v1beta1.AddressStringToBytesResponse"
> = {
  $type: "cosmos.auth.v1beta1.AddressStringToBytesResponse" as const,

  encode(message: AddressStringToBytesResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.addressBytes.length !== 0) {
      writer.uint32(10).bytes(message.addressBytes);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AddressStringToBytesResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAddressStringToBytesResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.addressBytes = reader.bytes();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AddressStringToBytesResponse {
    return { addressBytes: isSet(object.address_bytes) ? bytesFromBase64(object.address_bytes) : new Uint8Array(0) };
  },

  toJSON(message: AddressStringToBytesResponse): unknown {
    const obj: any = {};
    if (message.addressBytes.length !== 0) {
      obj.address_bytes = base64FromBytes(message.addressBytes);
    }
    return obj;
  },
  fromPartial(object: DeepPartial<AddressStringToBytesResponse>): AddressStringToBytesResponse {
    const message = createBaseAddressStringToBytesResponse();
    message.addressBytes = object.addressBytes ?? new Uint8Array(0);
    return message;
  },
};